+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Upcoming
+* Changed the rendering of stamina overlay
+  * New texture
+  * Fixes compatibility with Mantle hearts display
+
 ## 1.2.2
 * Fixed max Vigour modifier level
 * Hopefully fixed stamina overlay going over health bar